fix: 修复四个高优先级问题
行号溢出：
- LogEntry.line_no: u32 → u64，消除 adb 长期运行时的截断风险

bookmarks 隐含不变式：
- FilterSpec::matches() 新增 entry_idx: u32 参数
- 删除 line_no - 1 间接查找，直接用 entry_idx 查 bookmarks 集合
- spawn_filter_thread 传 j as u32

open_file TOCTOU：
- open_file() 打开文件后保留句柄，直接传给后台读取线程
- io.rs send_utf8_lines/send_decoded_lines 改为接受 File 而非 &Path
- 消除双重 open 和文件被删时静默无反馈的问题

save_filtered 主线程阻塞：
- 选路径对话框仍在主线程（egui 要求），选完后 spawn 后台线程写文件
- App 新增 save_result_rx，每帧 ui() 开头 try_recv() 收结果并更新状态栏
- 百万行日志保存时 UI 不再冻结

+++ src/model.rs
@@ -71,7 +71,7 @@ type Span = (u32, u32);
 #[derive(Debug, Clone)]
 pub struct LogEntry {
     raw: Box<str>,
-    pub line_no: u32,
+    pub line_no: u64,
     pub level: LevelMask,
     date: Span,
     time: Span,
@@ -93,7 +93,7 @@ impl LogEntry {
         tag: Span,
         message: Span,
     ) -> Self {
-        Self { raw, line_no: 0, level, date, time, pid, tid, tag, message }
+        Self { raw, line_no: 0u64, level, date, time, pid, tid, tag, message }
     }
 
     #[inline]
@@ -128,7 +128,7 @@ impl LogEntry {
         let i = push(&mut raw, tid);
         let g = push(&mut raw, tag);
         let m = push(&mut raw, message);
-        Self { raw: raw.into_boxed_str(), line_no: 1, level, date: d, time: t, pid: p, tid: i, tag: g, message: m }
+        Self { raw: raw.into_boxed_str(), line_no: 1u64, level, date: d, time: t, pid: p, tid: i, tag: g, message: m }
     }
 }
 
@@ -168,9 +168,9 @@ impl Model {
     }
 
     pub fn append(&mut self, mut entry: LogEntry) {
-        entry.line_no = (self.entries.len() as u32) + 1;
+        entry.line_no = (self.entries.len() as u64) + 1;
         if entry.level.contains(LevelMask::E) || entry.level.contains(LevelMask::F) {
-            self.error_lines.push(entry.line_no - 1);
+            self.error_lines.push(entry.line_no as u32 - 1);
         }
         // Bump per-value counts. Clone the key only when it's a new distinct
         // value (cardinality is tiny), not once per line.
